ci: address harness linter review feedback
