+++ .github/scripts/lint-harness-files.py
@@ -49,30 +49,7 @@ LINE_LIMIT_RULES = {
 }
 
 EMOJI_RANGES = (
-    (0x00A9, 0x00A9),
-    (0x00AE, 0x00AE),
-    (0x203C, 0x203C),
-    (0x2049, 0x2049),
-    (0x2122, 0x2122),
-    (0x2139, 0x2139),
-    (0x2194, 0x21AA),
-    (0x231A, 0x231B),
-    (0x2328, 0x2328),
-    (0x23CF, 0x23CF),
-    (0x23E9, 0x23F3),
-    (0x23F8, 0x23FA),
-    (0x24C2, 0x24C2),
-    (0x25AA, 0x25AB),
-    (0x25B6, 0x25B6),
-    (0x25C0, 0x25C0),
-    (0x25FB, 0x25FE),
     (0x2600, 0x27BF),
-    (0x2934, 0x2935),
-    (0x2B05, 0x2B55),
-    (0x3030, 0x3030),
-    (0x303D, 0x303D),
-    (0x3297, 0x3297),
-    (0x3299, 0x3299),
     (0x1F000, 0x1FAFF),
 )
